\name{as.big.matrix-methods}
\docType{methods}
\alias{as.big.matrix-methods}
\alias{as.big.matrix,matrix-method}
\alias{as.big.matrix,vector-method}
\title{Create a ``big.matrix'' from a matrix or vector.}
\description{
Create a \code{\link{big.matrix}} from a \code{matrix} or \code{vector};
a \code{vector} will result in a \code{\link{big.matrix}} with one column.
}
\section{Methods}{
\describe{
\item{\code{signature(x = "matrix")}}{ ... }
\item{\code{signature(x = "vector")}}{ ... }
}}
\keyword{methods}
