% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigmemory.R
\docType{methods}
\name{as.matrix,big.matrix-method}
\alias{as.matrix,big.matrix-method}
\title{Convert to base R matrix}
\usage{
\S4method{as.matrix}{big.matrix}(x)
}
\arguments{
\item{x}{A big.matrix object}
}
\description{
Extract values from a \code{big.matrix} object
and convert to a base R matrix object
}

