% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigmemory.R
\docType{methods}
\name{dim,big.matrix-method}
\alias{dim,big.matrix-method}
\title{Dimensions of a big.matrix object}
\usage{
\S4method{dim}{big.matrix}(x)
}
\arguments{
\item{x}{A \code{big.matrix} object}
}
\description{
Retrieve the dimensions of a \code{big.matrix} object
}

