% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bigmemory.R
\docType{methods}
\name{dimnames,big.matrix-method}
\alias{dimnames,big.matrix-method}
\alias{dimnames<-,big.matrix,list-method}
\title{Dimnames of a big.matrix Object}
\usage{
\S4method{dimnames}{big.matrix}(x)

\S4method{dimnames}{big.matrix,list}(x) <- value
}
\arguments{
\item{x}{A big.matrix object}

\item{value}{A possible value for \code{dimnames(x)}}
}
\description{
Retrieve or set the dimnames of an object
}

