% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigmemory.R
\docType{methods}
\name{Extract,big.matrix}
\alias{Extract,big.matrix}
\alias{[,big.matrix,ANY,ANY,missing-method}
\alias{[<-,big.matrix,ANY,ANY,ANY-method}
\alias{[,big.matrix,ANY,ANY,logical-method}
\alias{[,big.matrix,missing,ANY,missing-method}
\alias{[,big.matrix,missing,ANY,logical-method}
\alias{[,big.matrix,ANY,missing,missing-method}
\alias{[,big.matrix,ANY,missing,logical-method}
\alias{[,big.matrix,missing,missing,missing-method}
\alias{[,big.matrix,missing,missing,logical-method}
\alias{[,big.matrix,matrix,missing,missing-method}
\alias{[<-,big.matrix,numeric,numeric,ANY-method}
\alias{[<-,big.matrix,numeric,logical,ANY-method}
\alias{[<-,big.matrix,logical,numeric,ANY-method}
\alias{[<-,big.matrix,logical,logical,ANY-method}
\alias{[<-,big.matrix,logical,character,ANY-method}
\alias{[<-,big.matrix,numeric,character,ANY-method}
\alias{[<-,big.matrix,missing,missing,ANY-method}
\alias{[<-,big.matrix,missing,numeric,ANY-method}
\alias{[<-,big.matrix,missing,logical,ANY-method}
\alias{[<-,big.matrix,numeric,missing,numeric-method}
\alias{[<-,big.matrix,logical,missing,numeric-method}
\alias{[<-,big.matrix,numeric,missing,matrix-method}
\alias{[<-,big.matrix,logical,missing,matrix-method}
\alias{[<-,big.matrix,character,character,ANY-method}
\alias{[<-,big.matrix,missing,character,ANY-method}
\alias{[<-,big.matrix,character,missing,ANY-method}
\alias{[<-,big.matrix,missing,missing,numeric-method}
\alias{[<-,big.matrix,matrix,missing,numeric-method}
\title{Extract or Replace}
\usage{
\S4method{[}{big.matrix,ANY,ANY,missing}(x, i, j, drop)

\S4method{[}{big.matrix,ANY,ANY,logical}(x, i, j, drop)

\S4method{[}{big.matrix,missing,ANY,missing}(x, i, j, drop)

\S4method{[}{big.matrix,missing,ANY,logical}(x, i, j, drop)

\S4method{[}{big.matrix,ANY,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{big.matrix,ANY,missing,logical}(x, i, j, drop)

\S4method{[}{big.matrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{big.matrix,missing,missing,logical}(x, i, j, drop)

\S4method{[}{big.matrix,matrix,missing,missing}(x, i, j, drop)

\S4method{[}{big.matrix,numeric,numeric,ANY}(x, i, j) <- value

\S4method{[}{big.matrix,numeric,logical,ANY}(x, i, j) <- value

\S4method{[}{big.matrix,logical,numeric,ANY}(x, i, j) <- value

\S4method{[}{big.matrix,logical,logical,ANY}(x, i, j) <- value

\S4method{[}{big.matrix,logical,character,ANY}(x, i, j) <- value

\S4method{[}{big.matrix,numeric,character,ANY}(x, i, j) <- value

\S4method{[}{big.matrix,missing,missing,ANY}(x, i, j) <- value

\S4method{[}{big.matrix,missing,numeric,ANY}(x, i, j) <- value

\S4method{[}{big.matrix,missing,logical,ANY}(x, i, j) <- value

\S4method{[}{big.matrix,numeric,missing,numeric}(x, i, j, ...) <- value

\S4method{[}{big.matrix,logical,missing,numeric}(x, i, j, ...) <- value

\S4method{[}{big.matrix,numeric,missing,matrix}(x, i, j, ...) <- value

\S4method{[}{big.matrix,logical,missing,matrix}(x, i, j, ...) <- value

\S4method{[}{big.matrix,character,character,ANY}(x, i, j) <- value

\S4method{[}{big.matrix,missing,character,ANY}(x, j) <- value

\S4method{[}{big.matrix,character,missing,ANY}(x, i) <- value

\S4method{[}{big.matrix,missing,missing,numeric}(x, i, j) <- value

\S4method{[}{big.matrix,matrix,missing,numeric}(x, i, j) <- value
}
\arguments{
\item{x}{A \code{big.matrix object}}

\item{i}{Indices specifying the rows}

\item{j}{Indices specifying the columns}

\item{drop}{Logical indication if reduce to minimum dimensions}

\item{...}{Additional arguments}

\item{value}{typically an array-like R object of similar class}
}
\description{
Extract or replace big.matrix elements
}
