% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetMatrixSize}
\alias{GetMatrixSize}
\title{big.matrix size}
\usage{
GetMatrixSize(bigMat)
}
\arguments{
\item{bigMat}{a \code{big.matrix} object}
}
\description{
Returns the size of the created matrix in bytes
}

