% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigmemory.R
\docType{methods}
\name{head,big.matrix-method}
\alias{head,big.matrix-method}
\alias{tail,big.matrix-method}
\title{Return First or Last Part of a big.matrix Object}
\usage{
\S4method{head}{big.matrix}(x, n = 6)

\S4method{tail}{big.matrix}(x, n = 6)
}
\arguments{
\item{x}{A big.matrix object}

\item{n}{A single integer for the number of rows to return}
}
\description{
Returns the first or last parts of a \code{big.matrix}
object.
}
