% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigmemory.R
\name{is.float,numeric-method}
\alias{is.float,numeric-method}
\title{Is Float?}
\usage{
\S4method{is.float}{numeric}(x)
}
\arguments{
\item{x}{A numeric value}
}
\description{
Check if R numeric value has float flag
}
