% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigmemory.R
\name{is.float}
\alias{is.float}
\title{Check if Float}
\usage{
is.float(x)
}
\arguments{
\item{x}{An object to be evaluated if float}
}
\description{
Check to see if the elements of a big.matrix object are floats.
}
