% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigmemory.R
\docType{methods}
\name{length,big.matrix-method}
\alias{length,big.matrix-method}
\title{Length of a big.matrix object}
\usage{
\S4method{length}{big.matrix}(x)
}
\arguments{
\item{x}{A \code{big.matrix} object}
}
\description{
Get the length of a \code{big.matrix} object
}

