% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigmemory.R
\docType{methods}
\name{ncol,big.matrix-method}
\alias{ncol,big.matrix-method}
\alias{nrow,big.matrix-method}
\title{The Number of Rows/Columns of a big.matrix}
\usage{
\S4method{ncol}{big.matrix}(x)

\S4method{nrow}{big.matrix}(x)
}
\arguments{
\item{x}{A big.matrix object}
}
\value{
An integer of length 1
}
\description{
\code{nrow} and \code{ncol} return the number of
rows or columns present in a \code{big.matrix} object.
}
