% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\docType{methods}
\name{transpose,big.matrix-method}
\alias{transpose,big.matrix-method}
\title{Transposition}
\usage{
\S4method{transpose}{big.matrix}(x, ...)
}
\arguments{
\item{x}{A \link[bigmemory:big.matrix-class]{big.matrix}.}

\item{...}{Arguments passed on to \code{big.matrix}
\describe{
  \item{separated}{use separated column organization of the data; 
see details.}
  \item{backingfile}{the root name for the file(s) for the cache of \code{x}.}
  \item{backingpath}{the path to the directory containing the file 
backing cache.}
  \item{descriptorfile}{the name of the file to hold the backingfile 
description, for subsequent use with \code{\link{attach.big.matrix}}; 
if \code{NULL}, the \code{backingfile} is used as the root part of the 
descriptor file name.  The descriptor file is placed in the same directory 
as the backing files.}
  \item{binarydescriptor}{the flag to specify if the binary RDS format 
should be used for the backingfile description, for subsequent use with 
\code{\link{attach.big.matrix}}; if \code{NULL} of \code{FALSE}, the 
\code{dput()} file format is used.}
  \item{shared}{\code{TRUE} by default, and always \code{TRUE} if the 
\code{big.matrix} is file-backed.  For a non-filebacked \code{big.matrix}, 
\code{shared=FALSE} uses non-shared memory, which can be more stable for 
large (say, >50% of RAM) objects.  Shared memory allocation can sometimes 
fail in such cases due to exhausted shared-memory resources in the system.}
}}
}
\value{
The transposed \code{big.matrix}. Its dimension, type and dimnames
are automatically determined from the input \code{big.matrix}.
}
\description{
This function implements a simple cache-oblivious
algorithm for the transposition of a \code{big.matrix}.
}
\examples{
N <- 7
M <- 11
X <- big.matrix(N, M, dimnames = list(letters[1:N], LETTERS[1:M]))
X[] <- rnorm(length(X))

Xt <- transpose(X)
print(Xt[,])

print(identical(
  t(X[,]),
  Xt[,]
))
}
