% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigmemory.R
\docType{methods}
\name{typeof,big.matrix-method}
\alias{typeof,big.matrix-method}
\title{The Type of a big.matrix Object}
\usage{
\S4method{typeof}{big.matrix}(x)
}
\arguments{
\item{x}{A \code{big.matrix} object}
}
\description{
\code{typeof} returns the storage type of a
\code{big.matrix} object
}
