% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bigmemory.R
\docType{methods}
\name{Extract,big.matrix}
\alias{Extract,big.matrix}
\alias{[,big.matrix,ANY,ANY,logical-method}
\alias{[,big.matrix,ANY,ANY,missing-method}
\alias{[,big.matrix,ANY,missing,logical-method}
\alias{[,big.matrix,ANY,missing,missing-method}
\alias{[,big.matrix,matrix,missing,missing-method}
\alias{[,big.matrix,missing,ANY,logical-method}
\alias{[,big.matrix,missing,ANY,missing-method}
\alias{[,big.matrix,missing,missing,logical-method}
\alias{[,big.matrix,missing,missing,missing-method}
\alias{[<-,big.matrix,ANY,ANY-method}
\alias{[<-,big.matrix,ANY,missing-method}
\alias{[<-,big.matrix,matrix,missing-method}
\alias{[<-,big.matrix,missing,ANY-method}
\alias{[<-,big.matrix,missing,missing-method}
\title{Extract or Replace big.matrix elements}
\usage{
\S4method{[}{big.matrix,ANY,ANY,missing}(x, i, j, drop)

\S4method{[}{big.matrix,ANY,ANY,logical}(x, i, j, drop)

\S4method{[}{big.matrix,missing,ANY,missing}(x, i, j, drop)

\S4method{[}{big.matrix,missing,ANY,logical}(x, i, j, drop)

\S4method{[}{big.matrix,ANY,missing,missing}(x, i, j, drop)

\S4method{[}{big.matrix,ANY,missing,logical}(x, i, j, drop)

\S4method{[}{big.matrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{big.matrix,missing,missing,logical}(x, i, j, drop)

\S4method{[}{big.matrix,matrix,missing,missing}(x, i, j, drop)

\S4method{[}{big.matrix,ANY,ANY}(x, i, j) <- value

\S4method{[}{big.matrix,missing,ANY}(x, i, j) <- value

\S4method{[}{big.matrix,ANY,missing}(x, i, j) <- value

\S4method{[}{big.matrix,missing,missing}(x, i, j) <- value

\S4method{[}{big.matrix,matrix,missing}(x, i, j) <- value
}
\arguments{
\item{x}{A \code{big.matrix object}}

\item{i}{Indices specifying the rows}

\item{j}{Indices specifying the columns}

\item{drop}{Logical indication if reduce to minimum dimensions}

\item{value}{typically an array-like R object of similar class}
}
\description{
Extract or Replace big.matrix elements
}

