% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigmemory.R
\docType{methods}
\name{mwhich-methods}
\alias{mwhich-methods}
\alias{mwhich,big.matrix,ANY,ANY,ANY,character-method}
\alias{mwhich,big.matrix,ANY,ANY,ANY,missing-method}
\alias{mwhich,matrix,ANY,ANY,ANY,character-method}
\alias{mwhich,matrix,ANY,ANY,ANY,missing-method}
\title{Expanded ``which''-like functionality.}
\description{
Implements \code{\link{which}}-like functionality for a
\code{\link{big.matrix}}, with additional options for efficient comparisons
(executed in \acronym{C++}); also works for regular numeric matrices without
the memory overhead.
test
}
\section{Methods}{
 
\describe{ 
 \item{signature(x = "big.matrix=", cols
 = "ANY", vals = "ANY",", " comps = "ANY", op = "character")}{
 ... } 
 \item{signature(x = "big.matrix", cols = "ANY", vals =
 "ANY",", " comps = "ANY", op = "missing")}{ ... }
 \item{signature(x = "matrix", cols = "ANY", vals = "ANY",", "
 comps = "ANY", op = "character")}{ ... } 
 \item{signature(x = "matrix", cols = "ANY", vals = "ANY",", 
 " comps = "ANY", op = "missing")}{ ... } }
}

\seealso{
\code{\link{big.matrix}}, \code{\link{which}}, \code{\link{mwhich}}
}
\keyword{methods}
